fix disconnect deadlock
bufferPayload used to grab inFlightBufferLock at the beginning. if either token decoding or payload marshalling failed, Disconnect would be unable to grab it and deadlock. this patch delays grabbing the lock until we know we want to begin mutating the buffer. deferred unlocking of inFlightBufferLock to try and prevent future deadlock issues.

i also refactored bufferPayload to stop disconnecting and instead return errors, since i don't see a reason why we would disconnect our APNS socket due to a bad message that is never sent. we now log the bad message and continue to the next payload.